+++ builtin/logical/pki/path_acme_directory.go
@@ -1,12 +1,9 @@
 package pki
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"net/url"
-	"strings"
 
 	"github.com/hashicorp/vault/sdk/framework"
 	"github.com/hashicorp/vault/sdk/logical"
@@ -52,71 +49,6 @@ func patternAcmeDirectory(b *backend, pattern string) *framework.Path {
 
 		HelpSynopsis:    pathAcmeDirectoryHelpSync,
 		HelpDescription: pathAcmeDirectoryHelpDesc,
-	}
-}
-
-type acmeOperation func(acmeCtx *acmeContext, r *logical.Request, _ *framework.FieldData) (*logical.Response, error)
-
-type acmeContext struct {
-	baseUrl *url.URL
-	sc      *storageContext
-}
-
-func (b *backend) acmeWrapper(op acmeOperation) framework.OperationFunc {
-	return acmeErrorWrapper(func(ctx context.Context, r *logical.Request, data *framework.FieldData) (*logical.Response, error) {
-		sc := b.makeStorageContext(ctx, r.Storage)
-
-		if false {
-			// TODO sclark: Check if ACME is enable here
-			return nil, fmt.Errorf("ACME is disabled in configuration: %w", ErrServerInternal)
-		}
-
-		baseUrl, err := getAcmeBaseUrl(sc, r.Path)
-		if err != nil {
-			return nil, err
-		}
-
-		acmeCtx := &acmeContext{
-			baseUrl: baseUrl,
-			sc:      sc,
-		}
-
-		return op(acmeCtx, r, data)
-	})
-}
-
-func getAcmeBaseUrl(sc *storageContext, path string) (*url.URL, error) {
-	cfg, err := sc.getClusterConfig()
-	if err != nil {
-		return nil, fmt.Errorf("failed loading cluster config: %w", err)
-	}
-
-	if cfg.Path == "" {
-		return nil, fmt.Errorf("ACME feature requires local cluster path configuration to be set: %w", ErrServerInternal)
-	}
-
-	baseUrl, err := url.Parse(cfg.Path)
-	if err != nil {
-		return nil, fmt.Errorf("ACME feature a proper URL configured in local cluster path: %w", ErrServerInternal)
-	}
-
-	directoryPrefix := ""
-	lastIndex := strings.LastIndex(path, "/acme/")
-	if lastIndex != -1 {
-		directoryPrefix = path[0:lastIndex]
-	}
-
-	return baseUrl.JoinPath(directoryPrefix, "/acme/"), nil
-}
-
-func acmeErrorWrapper(op framework.OperationFunc) framework.OperationFunc {
-	return func(ctx context.Context, r *logical.Request, data *framework.FieldData) (*logical.Response, error) {
-		resp, err := op(ctx, r, data)
-		if err != nil {
-			return TranslateError(err)
-		}
-
-		return resp, nil
 	}
 }
